Refactor RISC-V and AArch64 boot code to integrate Limine support and remove unused components
