+++ kernel/Cargo.toml
@@ -17,9 +17,10 @@ limine = "0.5"
 font8x8 = { version = "0.3", default-features = false, features = ["unicode"] }
 
 [features]
-default = ["network", "user-fpu", "user-vector", "hypervisor"]
+default = ["network", "user-fpu", "user-vector", "hypervisor", "limine"]
 profiler = ["dep:lazy_static"]
 network = []
 user-fpu = []
 user-vector = []
 hypervisor = []
+limine = []
