+++ kernel/src/arch/riscv64/boot/mod.rs
@@ -1,3 +1,6 @@
+#[cfg(feature = "limine")]
+pub mod limine;
+
 use core::arch::asm;
 
 use crate::{
