+++ kernel/src/boot/mod.rs
@@ -1,1 +1,2 @@
+#[cfg(feature = "limine")]
 pub mod limine;
