+++ kernel/targets/riscv64gc-unknown-none-elf.json
@@ -13,7 +13,7 @@
   "llvm-target": "riscv64",
   "max-atomic-width": 64,
   "metadata": {
-    "description": "Bare RISC-V (RV64IMAFDC ISA)",
+    "description": "Bare RISC-V (RV64IMACH ISA)",
     "host_tools": false,
     "std": false,
     "tier": 2
